requirements.txt
streamlit
langchain
langchain-openai
langchain-experimental
python-dotenv
matplotlib
seaborn
openai
tabulate
